+++ Suyeon/code/[vscode] NaverNews-(1)Collecting.py
@@ -61,7 +61,7 @@ def news_list(keyword, startdate, enddate):
     return pd.DataFrame(li, columns=['title', 'date', 'media', 'content', 'url'])
 
 # 크롤링할 데이터 (키워드, 시작날짜, 종료날짜)
-result_df = news_list('테슬라', '2022.04.01', '2022.04.30')
+result_df = news_list('테슬라', '2022.05.01', '2022.05.31')
 
 # 크롤링 데이터 CSV파일로 저장
-result_df.to_csv('[2022-Apr]news_data.csv', index=False, encoding='utf-8-sig')
+result_df.to_csv('[2022-May]news_data.csv', index=False, encoding='utf-8-sig')
